LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/require-corp-cached-images.https-expected.txt

CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/corp-image.py because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Cannot load image https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/corp-image.py due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/corp-image.py because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Cannot load image https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/corp-image.py due to access control checks.


PASS NETWORK-https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/corp-image.py
PASS NETWORK-https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/corp-image.py?corp-cross-origin=1
PASS CACHED-https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/corp-image.py
PASS CACHED-https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/corp-image.py?corp-cross-origin=1
PASS main_test
